# Directly store the sparse matrix in the SCS format