--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -22,5 +22,5 @@ end
 
 include("options.jl")
 
-include("MPB_wrapper.jl")
 include("MOI_wrapper.jl")
+include("MPB_wrapper.jl")