Closes https://github.com/jump-dev/SCS.jl/issues/186