@kalmarek 2020-07-26
I can't really comment on the MOI part, but changes in `c_wrapper.jl` look ok. Also there are some lost `@show`s in `types.jl` ;)