@blegat 2020-09-03
Yes, let's merge so that it's used in SolverTests